area

+++ khokhryakova.daria/T3/commands.cpp
@@ -159,7 +159,7 @@ void khokhryakova::area(std::istream& in, std::ostream& out, const std::vector< 
   subcommands["EVEN"] = std::bind(areaEven, std::cref(polygons));
   subcommands["ODD"] = std::bind(areaOdd, std::cref(polygons));
   subcommands["MEAN"] = std::bind(areaMedian, std::cref(polygons));
-  size_t size = 0;
+  double size = 0;
   try
   {
     size = subcommands.at(command)();
